[frontend] update bash script for build

+++ run.sh
@@ -4,8 +4,9 @@ set -u
 
 __bash_dir__="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-SKIP_SDC_SOURCE_TEST="${DOWNLOAD_SDC:-false}"
-SKIP_BUILD_LIB="${DOWNLOAD_SDC:-true}"
+# SKIP_SDC_SOURCE_TEST="${SKIP_SDC_SOURCE_TEST:-false}"
+SKIP_BUILD_LIB="${SKIP_BUILD_LIB:-false}"
+SKIP_RUN_SDC="${SKIP_RUN_SDC:-false}"
 DEV_MODE="${DEV_MODE:-true}"
 
 make_dist_dir() {
@@ -25,7 +26,7 @@ download_sdc() {
         curl -O http://nightly.streamsets.com.s3-us-west-2.amazonaws.com/datacollector/latest/tarball/streamsets-datacollector-core-${SDC_VERSION}.tgz
     fi
     rm -rf "${__bash_dir__}/dist/sdc"
-    tar -xvf streamsets-datacollector-core-${SDC_VERSION}.tgz 
+    tar -xvf streamsets-datacollector-core-${SDC_VERSION}.tgz  &> /dev/null
     mv streamsets-datacollector-${SDC_VERSION} sdc
     rm  -rf sdc/sdc-static-web
     cd "${__bash_dir__}"
@@ -33,43 +34,59 @@ download_sdc() {
 
 install_ui_lib() {
     cd "${__bash_dir__}/datacollector-ui"
-    yarn install
-    yarn install -g bower
-    yarn install -g grunt-cli
-    bower install
+    if hash yarn 2>/dev/null; then
+        yarn install
+        yarn global add bower
+        yarn global add grunt-cli
+        bower install
+    else
+        echo 'please install yarn for build environment.'
+        exit 1;
+    fi
+
+    
 }
 build_ui() {
     echo "start to build html"
     cd "${__bash_dir__}/datacollector-ui"
   
-    grunt build
+    grunt build --force
 }
 
 watch_ui() {
     echo "start to grunt watch"
     cd "${__bash_dir__}/datacollector-ui"
-    grunt watch
+    grunt watch --force
 }
 
 run_sdc() {
+    echo "start to check sdc download"
+    download_sdc
+    
     echo "start to run sdc"
     cd "${__bash_dir__}"
+    
     export SDC_FILE_LIMIT=1024
     # dist/sdc/bin/streamsets dc
     BUILD_ID=dontKillMe  nohup dist/sdc/bin/streamsets dc &
 }
 
 main () {
-
-    make_dist_dir
     
-    if [ "$SKIP_SDC_SOURCE_TEST" <> "true" ]; then
-        download_sdc
-    fi
-    kill $(lsof -t -i:18630)
-    run_sdc
+    make_dist_dir 
+    
+    # if [ "$SKIP_SDC_SOURCE_TEST" != "true" ]; then
+    #     download_sdc
+    # fi
 
-    if [ "$SKIP_BUILD_LIB" <> "true" ]; then
+    if [ "$SKIP_RUN_SDC" != "true" ]; then
+        kill $(lsof -t -i:18630)
+        echo 'starting sdc'
+        run_sdc
+    fi
+    
+
+    if [ "$SKIP_BUILD_LIB" != "true" ]; then
        install_ui_lib
     fi
 
@@ -77,8 +94,8 @@ main () {
         watch_ui
     else 
         build_ui
+        echo 'Done: built dist html files in ./dist/sdc/sdc-static-web'
     fi 
-    
     
 }
 main "$@"
